Test incremental upload of tab delim. data types from umbrella script
- Split big tab. delim test to multiple tests based on data type.

- Use ImportProfileData instead of ImportTabDelimData for testing.
  - We cover more logic with such tests.
  - This is more stable interface. ImportTabDelimData can be refactored.
