+++ src/test/resources/incremental/copy_number_alteration/data_cna_discrete.txt
