+++ src/test/resources/incremental/generic_assay/meta_treatment_ic50.txt
@@ -1,0 +1,12 @@
+cancer_study_identifier: study_tcga_pub
+genetic_alteration_type: GENERIC_ASSAY
+generic_assay_type: TREATMENT_RESPONSE
+datatype: LIMIT-VALUE
+stable_id: treatment_ic50
+profile_name: IC50 values of compounds on cellular phenotype readout
+profile_description: IC50 (compound concentration resulting in half maximal inhibition) of compounds on cellular phenotype readout of cultured mutant cell lines.
+data_filename: data_treatment_ic50.txt
+show_profile_in_analysis_tab: true
+pivot_threshold_value: 0.1
+value_sort_order: ASC
+generic_entity_meta_properties: NAME,DESCRIPTION,URL
