+++ src/test/resources/incremental/mrna_expression/meta_expression_Zscores.txt
@@ -1,0 +1,8 @@
+cancer_study_identifier: study_tcga_pub
+genetic_alteration_type: MRNA_EXPRESSION
+datatype: Z-SCORE
+stable_id: mrna
+profile_description: Expression levels (Agilent microarray).
+show_profile_in_analysis_tab: false
+profile_name: mRNA expression (microarray)
+data_filename: data_expression_Zscores.txt
