+++ src/test/resources/incremental/protein_level/meta_rppa.txt
@@ -1,0 +1,7 @@
+cancer_study_identifier: study_tcga_pub
+genetic_alteration_type: PROTEIN_LEVEL
+datatype: LOG2-VALUE
+stable_id: rppa
+profile_name: Test RPPA
+profile_description: Test protein level data
+data_filename: data_rppa.txt
